+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # Changelog
 
 ## vTBD
-- [BP-1785](https://movai.atlassian.net/browse/BP-1785): Flow validations carry recursive context for parameter parsing validations.
-  - Project validation no longer includes parameter parsing validations.
+- [BP-1785](https://movai.atlassian.net/browse/BP-1785): Project and Flow validations carry recursive context for parameter parsing validations.
 
 ## v3.30.2
 - [BP-1785](https://movai.atlassian.net/browse/BP-1785): Links should allow connection of combinations of ROS1/ROS2 Publisher/Subscriber.